==> scripts/3_audio_features_query.ipynb <==
# Audio Features Query

Based on: https://stmorse.github.io/journal/spotify-api.html

import requests
import pandas as pd
import numpy as np
from tqdm.notebook import tqdm # added for progress bars
import time
import csv 

client_id = 'YOUR CLIENT ID'
client_secret = 'YOUR CLIENT SECRET'

AUTH_URL = 'https://accounts.spotify.com/api/token'

# POST
auth_response = requests.post(AUTH_URL, {
    'grant_type': 'client_credentials',
    'client_id': client_id,
    'client_secret': client_secret,
})

# convert the response to JSON
auth_response_data = auth_response.json()

# save the access token
access_token = auth_response_data['access_token']

headers = {
    'Authorization': 'Bearer {token}'.format(token = access_token)
}

# base URL of all Spotify API endpoints
BASE_URL = 'https://api.spotify.com/v1/'

# Track ID from the URI
track_id = '21GtSB7nXJUinRaxOvClgX'

# actual GET request with proper header
r = requests.get(BASE_URL + 'audio-features/' + track_id, headers = headers)

r = r.json()
r

Now import and interate over all song ids to query the audio features and append them to the IDs (song ID array was made in R.) There are two ways to query the data either song-by-song or in chunks of songs. The latter is more efficient as it decreases the number of requests.

# set file path
file_path = '/Users/sebastian/Documents/Uni/Sheffield (MSc)/2. Semester/Data Analysis and Viz/spotify_audio_feature_analysis/'

"Get Audio Features for *a* Track"

track_ids = pd.read_csv(file_path + 'data/4_na_songids.csv')
track_ids = track_ids['x'] # in case the there's an index variable included
track_ids

def get_audio_features(track_ids):
    data = []
    for ID in tqdm(track_ids):
        # get audio features (key, liveness, danceability, ect.)
        f = requests.get(BASE_URL + 'audio-features/' + ID, headers = headers)
        f = f.json()
        data.append(f)
        print(data)
    return(data)

"Get Audio Features for *Several* Tracks"

track_ids_chunk = pd.read_csv(file_path + 'data/2_songids_chunks.csv')
track_ids_chunk = track_ids_chunk['x']  # in case the there's an index variable included

def get_audio_features_several_tracks(track_id_array):
    data = []
    for ID in tqdm(track_id_array):
        # get audio features (key, liveness, danceability, ect.)
        f = requests.get(BASE_URL + 'audio-features?ids=' + ID, headers = headers) # several tracks 'audio-features?ids='
        f = f.json()
        data.append(f)
    return(data)

def transform_list(chunks):
    tracks = []
    for chunk in chunks:
        raw_chunk = chunk['audio_features']
        
        for track in raw_chunk:
            tracks.append(track)
        
    return(tracks)

df_chunk = get_audio_features_several_tracks(track_ids_chunk)

tracks = transform_list(df_chunk)

# there seem to be none values in list, check where they are
index_none = [i for i in range(len(tracks)) if tracks[i] == None] 

index_none

# remove entrys with none values from list
tracks = [i for i in tracks if i] 

# transform to data frame
final = pd.DataFrame(tracks)

# save to csv
final.to_csv(file_path + 'data/3_audio_features.csv")
